pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Md Badiuzzaman Pranto", email = "prantoamt@gmail.com"},
]
dependencies = [
    "Django==4.1.5",
    "djangorestframework==3.14",
    "markdown>=3.4.1",
    "django-filter>=22.1",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool.pdm.scripts]
startdev = "python3 assignment/manage.py runserver 0.0.0.0:8000"
makemigrations = "python3 assignment/manage.py makemigrations"
migrate = "python3 assignment/manage.py migrate"
startapp = "python3 assignment/manage.py startapp"
collectstatic = "python3 assignment/manage.py collectstatic"
fetchcompany = "python3 assignment/manage.py fetchcompany"
formatcode = "python3 -m black assignment"

[tool.pdm.dev-dependencies]
formatter = [
    "black>=22.12.0",
]
